+++ components/Cards/CardList.tsx
@@ -33,7 +33,7 @@ const CardList: React.FC<CardListProps> = ({ columnId }) => {
     setTagColors(newTagColors);
   }, [cards]); // cards가 변경될 때만 실행
 
-  console.log("selectedCard:", selectedCard);
+  //console.log("selectedCard:", selectedCard);
 
   useEffect(() => {
     const fetchCards = async () => {
